feat(TMG-3003): separate util functions into separate files
