+++ packages/save-and-share-bar/src/save-and-share-bar.js
@@ -10,7 +10,7 @@ import UserState from "@times-components/user-state";
 import { SectionContext } from "@times-components/context";
 import { SaveStar } from "@times-components/ts-components";
 
-import { checkForSymphonyExperiment } from "@times-components/utils";
+import checkForSymphonyExperiment from "@times-components/utils";
 import getTokenisedArticleUrlApi from "./get-tokenised-article-url-api";
 import withTrackEvents from "./tracking/with-track-events";
 import SharingApiUrls from "./constants";
