+++ packages/utils/src/get-cookie-value.js
@@ -1,0 +1,8 @@
+const getCookieValue = cookieName => {
+  const allCookies = document.cookie;
+  const cookiesArray = allCookies.split(";");
+  const targetCookie = cookiesArray.find(cookie => cookie.includes(cookieName));
+  return targetCookie ? targetCookie.split("=")[1] : null;
+};
+
+export default getCookieValue;
